==> attempt_03/04_assign_topics.ipynb <==
# Assign topic distributions to authors
* List all authors in the dataset.
* For each author, merge the abstracts of all their articles, apply pre-processing filters and lemmatization.
* Finally, apply the trained LDA topic model to extract the topic distribution from each text and assign this topic distribution to each author. 

import pandas as pd
import pickle
import os
import arxiv_utils
import tok

DATA_PATH = '../data'
MODELS_PATH = '../models'

## Load the article abstracts and metadata for the train/validate/test datasets

def load_df(name):
    return pd.read_csv(
        os.path.join(DATA_PATH, name), 
        index_col=0, 
        converters={"authors_parsed": lambda x:[entry.replace("\'", '').replace('"', '').strip("[]") for entry in x.split('", "')]})

train_df = load_df('arxiv_train.csv')
validate_df = load_df('arxiv_validate.csv')
test_df = load_df('arxiv_test.csv')

## List all authors in the dataset

def flatten(xss):
    """Flatten a list of lists"""
    return [x for xs in xss for x in xs]

def get_unique_authors(df):
    """Given a dataframe, return unique authors"""
    authors = flatten(df['authors_parsed'])
    return set(authors)

def count_authors(df):
    """Given a dataframe, return count of unique authors"""
    return len(get_unique_authors(df))

train_authors = arxiv_utils.get_unique_authors(train_df)
validate_authors = arxiv_utils.get_unique_authors(validate_df)
test_authors = arxiv_utils.get_unique_authors(test_df)
print(f"The train dataset has {len(train_authors)} unique authors, the validate dataset {len(validate_authors)}, the test dataset {len(test_authors)}.")

## For each author, merge the abstracts of all their articles

def get_abstracts_by_author(author, metadata_df):
    """
    Get the abstracts of all articles by this author.
    author: string
    metadata_df: pandas.DataFrame dataset
    returns: list<string>
    """
    idx = [author in authors_parsed for authors_parsed in metadata_df['authors_parsed']]
    abstracts = metadata_df[idx]['abstract']
    return list(abstracts)

def merge_abstracts_by_author(author_list, metadata_df):
    """
    Merge the abstracts of all articles by all authors.
    author_list: set<string> set of unique author names
    metadata_df: pandas.DataFrame dataset
    returns: list<string> merged abstracts
    """
    txts = []
    counter = 0
    for author in author_list:
        if counter % 25000 == 0:
            print(f"Processing author {counter}/{len(author_list)}")
        abstracts = get_abstracts_by_author(author, metadata_df)
        txt = ' '.join(abstracts)
        txts.append(txt)
        counter += 1
    return txts

%%time

print("Processing train dataset")
train_authors_txts = merge_abstracts_by_author(train_authors, train_df)
print("Processing validate dataset")
validate_authors_txts = merge_abstracts_by_author(validate_authors, validate_df)
print("Processing test dataset")
test_authors_txts = merge_abstracts_by_author(test_authors, test_df)

## Apply pre-processing filters and lemmatization to the texts for each author

with open(os.path.join(DATA_PATH, 'dictionary.pickle'), 'rb') as handle:
    dictionary = pickle.load(handle)

%%time

def tokenize_dataset(dictionary, txt_list):
    _texts = tok.clean(txt_list)
    return(tok.make_corpus(dictionary, _texts))

print("Tokenizing train dataset")
train_corpus_authors = tokenize_dataset(dictionary, train_authors_txts)
print("Tokenizing validate dataset")
validate_corpus_authors = tokenize_dataset(dictionary, validate_authors_txts)
print("Tokenizing test dataset")
test_corpus_authors = tokenize_dataset(dictionary, test_authors_txts)

## Apply the trained LDA topic model

Load the topic model
Load the LDA topic model fitted in [03_fit_topic_model](./03_fit_topic_model.ipynb)

with open(os.path.join(MODELS_PATH, 'topic_model9.pickle'), 'rb') as handle:
    topic_model9 = pickle.load(handle)

with open(os.path.join(MODELS_PATH, 'topic_model6.pickle'), 'rb') as handle:
    topic_model6 = pickle.load(handle)

with open(os.path.join(MODELS_PATH, 'topic_model12.pickle'), 'rb') as handle:
    topic_model12 = pickle.load(handle)

Extract the topic distribution from each text and assign this topic distribution to each author.

def get_topic_details(topic_model, corpus):
    """
    Returns a list of pandas Series object of tuples. 
    Each tuple is a topic number and the topic probability for this entry in the corpus.
    Example: 
        [[(0, 0.22764261), (4, 0.14444388), (5, 0.62411755)],
         [(1, 0.024827635), (2, 0.3290665), (3, 0.6061594), (5, 0.03431195)],
         [(0, 0.06239689), (3, 0.03924617), (5, 0.8926314)],
         [(3, 0.09784623), (5, 0.89414346)],...
        ...]
    If for a given entry, the topic's probability is 0, then the topic is not included in the Series for this entry.
    """
    topic_details_list = []
    for row in topic_model[corpus]:
        topic_details_list.append(row)
    return topic_details_list

def get_topic_dataframe(topic_model, corpus):
    """
    Returns a data frame with a column for each topic in the topic model.
    Each row stands for an entry in the corpus, each value for the probability of thos topic for this entry.
    If for a given entry, the topic's probability is 0, the the value in the entry's column corresponding to the topic is also 0.
    Example:
             	0 	1 	2 	3 	4 	5
        0 	0.227641 	0.000000 	0.000000 	0.000000 	0.144445 	0.624118
        1 	0.000000 	0.024817 	0.329062 	0.606161 	0.000000 	0.034325
        2 	0.062392 	0.000000 	0.000000 	0.039281 	0.000000 	0.892601
        3 	0.000000 	0.000000 	0.000000 	0.097728 	0.000000 	0.894262
    """
    topic_details = get_topic_details(topic_model, corpus)
    topics_entries = []  # topics for all entries
    num_topics = len(topic_model.get_topics())  # number of topics in the model
    for row in topic_details:
        topics_entry = [0] * num_topics
        for entry in row:  # all topic probabilities for this entry
            topic_num = entry[0]  # the topic number
            topic_prob = entry[1]  # the topic probability            
            topics_entry[topic_num] = topic_prob
        topics_entries.append(topics_entry)
    return pd.DataFrame(topics_entries, columns=range(0, num_topics))

%%time

train_topics9_authors = get_topic_dataframe(topic_model9, train_corpus_authors)
validate_topics9_authors = get_topic_dataframe(topic_model9, validate_corpus_authors)
test_topics9_authors = get_topic_dataframe(topic_model9, test_corpus_authors)

train_topics6_authors = get_topic_dataframe(topic_model6, train_corpus_authors)
validate_topics6_authors = get_topic_dataframe(topic_model6, validate_corpus_authors)
test_topics6_authors = get_topic_dataframe(topic_model6, test_corpus_authors)

train_topics12_authors = get_topic_dataframe(topic_model12, train_corpus_authors)
validate_topics12_authors = get_topic_dataframe(topic_model12, validate_corpus_authors)
test_topics12_authors = get_topic_dataframe(topic_model12, test_corpus_authors)

## Add author names and save

train_topics9_authors["author"] = list(train_authors)
validate_topics9_authors["author"] = list(validate_authors)
test_topics9_authors["author"] = list(test_authors)

train_topics6_authors["author"] = list(train_authors)
validate_topics6_authors["author"] = list(validate_authors)
test_topics6_authors["author"] = list(test_authors)

train_topics12_authors["author"] = list(train_authors)
validate_topics12_authors["author"] = list(validate_authors)
test_topics12_authors["author"] = list(test_authors)

train_topics9_authors.to_csv(os.path.join(DATA_PATH, 'train_topics9_authors.csv'))
validate_topics9_authors.to_csv(os.path.join(DATA_PATH, 'validate_topics9_authors.csv'))
test_topics9_authors.to_csv(os.path.join(DATA_PATH, 'test_topics9_authors.csv'))

train_topics6_authors.to_csv(os.path.join(DATA_PATH, 'train_topics6_authors.csv'))
validate_topics6_authors.to_csv(os.path.join(DATA_PATH, 'validate_topics6_authors.csv'))
test_topics6_authors.to_csv(os.path.join(DATA_PATH, 'test_topics6_authors.csv'))

train_topics12_authors.to_csv(os.path.join(DATA_PATH, 'train_topics12_authors.csv'))
validate_topics12_authors.to_csv(os.path.join(DATA_PATH, 'validate_topics12_authors.csv'))
test_topics12_authors.to_csv(os.path.join(DATA_PATH, 'test_topics12_authors.csv'))

# Assign topic distributions to articles

# Load the tokenized abstracts
def load_tokenized_dataset(file_name):
    path = os.path.join(DATA_PATH, file_name)
    with open(path, 'rb') as handle:
        tokenized_dataset = pickle.load(handle)
    return tokenized_dataset

corpus_test = load_tokenized_dataset("corpus_test.pickle")
corpus_validate = load_tokenized_dataset("corpus_validate.pickle")
corpus_train = load_tokenized_dataset("corpus_train.pickle")

topics9_test_df = get_topic_dataframe(topic_model9, corpus_test)
topics9_validate_df = get_topic_dataframe(topic_model9, corpus_validate)
topics9_train_df = get_topic_dataframe(topic_model9, corpus_train)

topics6_test_df = get_topic_dataframe(topic_model6, corpus_test)
topics6_validate_df = get_topic_dataframe(topic_model6, corpus_validate)
topics6_train_df = get_topic_dataframe(topic_model6, corpus_train)

topics12_test_df = get_topic_dataframe(topic_model12, corpus_test)
topics12_validate_df = get_topic_dataframe(topic_model12, corpus_validate)
topics12_train_df = get_topic_dataframe(topic_model12, corpus_train)

## Load the article metadata and merge with the topics into one data frame

def load_df(name):
    return pd.read_csv(
        os.path.join(DATA_PATH, name), 
        index_col=0, 
        converters={"authors_parsed": lambda x:[entry.replace("\'", '').replace('"', '').strip("[]") for entry in x.split('", "')]})

validate_df = load_df('arxiv_validate.csv')
test_df = load_df('arxiv_test.csv')
train_df = load_df('arxiv_train.csv')

num_topics = len(topic_model9.get_topics())
for l in [topics9_test_df, topics9_validate_df, topics9_train_df]:
        l.columns = ["topic9_%d"%n for n in range(num_topics)]

num_topics = len(topic_model6.get_topics())
for l in [topics6_test_df, topics6_validate_df, topics6_train_df]:
        l.columns = ["topic6_%d"%n for n in range(num_topics)]

num_topics = len(topic_model12.get_topics())
for l in [topics12_test_df, topics12_validate_df, topics12_train_df]:
        l.columns = ["topic12_%d"%n for n in range(num_topics)]

def merge_df(article_df, topics_df):
    merged = article_df.reset_index(drop=True).join(topics_df.reset_index(drop=True))
    return merged

test9_df = merge_df(test_df, topics9_test_df)
validate9_df = merge_df(validate_df, topics9_validate_df)
train9_df = merge_df(train_df, topics9_train_df)

test6_df = merge_df(test_df, topics6_test_df)
validate6_df = merge_df(validate_df, topics6_validate_df)
train6_df = merge_df(train_df, topics6_train_df)

test12_df = merge_df(test_df, topics12_test_df)
validate12_df = merge_df(validate_df, topics12_validate_df)
train12_df = merge_df(train_df, topics12_train_df)

train9_df.to_csv(os.path.join(DATA_PATH, 'train_topics9.csv'))
validate9_df.to_csv(os.path.join(DATA_PATH, 'validate_topics9.csv'))
test9_df.to_csv(os.path.join(DATA_PATH, 'test_topics9.csv'))

train6_df.to_csv(os.path.join(DATA_PATH, 'train_topics6.csv'))
validate6_df.to_csv(os.path.join(DATA_PATH, 'validate_topics6.csv'))
test6_df.to_csv(os.path.join(DATA_PATH, 'test_topics6.csv'))

train12_df.to_csv(os.path.join(DATA_PATH, 'train_topics12.csv'))
validate12_df.to_csv(os.path.join(DATA_PATH, 'validate_topics12.csv'))
test12_df.to_csv(os.path.join(DATA_PATH, 'test_topics12.csv'))
